refactor: re-org integs into single crate to exploit parallelism
Shaves almost 10s off the test time, and the output looks a bit
cleaner.
