--- modules/fundamental/tests/graph.rs
@@ -1,536 +1,0 @@
-use std::convert::TryInto;
-use std::str::FromStr;
-use test_context::test_context;
-use test_log::test;
-use trustify_common::db::Transactional;
-use trustify_common::hashing::Digests;
-use trustify_common::purl::Purl;
-use trustify_common::sbom::SbomLocator;
-use trustify_entity::relationship::Relationship;
-use trustify_module_fundamental::sbom::service::SbomService;
-use trustify_module_ingestor::graph::Graph;
-use trustify_test_context::TrustifyContext;
-
-#[test_context(TrustifyContext, skip_teardown)]
-#[test(tokio::test)]
-async fn ingest_sboms(ctx: TrustifyContext) -> Result<(), anyhow::Error> {
-    let db = ctx.db;
-    let system = Graph::new(db);
-
-    let sbom_v1 = system
-        .ingest_sbom(
-            ("source", "http://sbom.com/test.json"),
-            &Digests::digest("8"),
-            "a",
-            (),
-            Transactional::None,
-        )
-        .await?;
-    let sbom_v1_again = system
-        .ingest_sbom(
-            ("source", "http://sbom.com/test.json"),
-            &Digests::digest("8"),
-            "b",
-            (),
-            Transactional::None,
-        )
-        .await?;
-    let sbom_v2 = system
-        .ingest_sbom(
-            ("source", "http://sbom.com/test.json"),
-            &Digests::digest("9"),
-            "c",
-            (),
-            Transactional::None,
-        )
-        .await?;
-
-    let _other_sbom = system
-        .ingest_sbom(
-            ("source", "http://sbom.com/other.json"),
-            &Digests::digest("10"),
-            "d",
-            (),
-            Transactional::None,
-        )
-        .await?;
-
-    assert_eq!(sbom_v1.sbom.sbom_id, sbom_v1_again.sbom.sbom_id);
-
-    assert_ne!(sbom_v1.sbom.sbom_id, sbom_v2.sbom.sbom_id);
-    Ok(())
-}
-
-#[test_context(TrustifyContext, skip_teardown)]
-#[test(tokio::test)]
-async fn ingest_and_fetch_sboms_describing_purls(
-    ctx: TrustifyContext,
-) -> Result<(), anyhow::Error> {
-    let db = ctx.db;
-    let system = Graph::new(db);
-
-    let sbom_v1 = system
-        .ingest_sbom(
-            ("source", "http://sbom.com/test.json"),
-            &Digests::digest("8"),
-            "a",
-            (),
-            Transactional::None,
-        )
-        .await?;
-    let sbom_v2 = system
-        .ingest_sbom(
-            ("source", "http://sbom.com/test.json"),
-            &Digests::digest("9"),
-            "b",
-            (),
-            Transactional::None,
-        )
-        .await?;
-    let sbom_v3 = system
-        .ingest_sbom(
-            ("source", "http://sbom.com/test.json"),
-            &Digests::digest("10"),
-            "c",
-            (),
-            Transactional::None,
-        )
-        .await?;
-
-    sbom_v1
-        .ingest_describes_package(
-            "pkg://maven/io.quarkus/quarkus-core@1.2.3".try_into()?,
-            Transactional::None,
-        )
-        .await?;
-
-    sbom_v2
-        .ingest_describes_package(
-            "pkg://maven/io.quarkus/quarkus-core@1.2.3".try_into()?,
-            Transactional::None,
-        )
-        .await?;
-
-    sbom_v3
-        .ingest_describes_package(
-            "pkg://maven/io.quarkus/quarkus-core@1.9.3".try_into()?,
-            Transactional::None,
-        )
-        .await?;
-
-    let found = system
-        .locate_sboms(
-            SbomLocator::Purl("pkg://maven/io.quarkus/quarkus-core@1.2.3".try_into()?),
-            Transactional::None,
-        )
-        .await?;
-
-    assert_eq!(2, found.len());
-    assert!(found.contains(&sbom_v1));
-    assert!(found.contains(&sbom_v2));
-
-    Ok(())
-}
-
-#[test_context(TrustifyContext, skip_teardown)]
-#[test(tokio::test)]
-async fn ingest_and_locate_sboms_describing_cpes(
-    ctx: TrustifyContext,
-) -> Result<(), anyhow::Error> {
-    let db = ctx.db;
-    let system = Graph::new(db);
-
-    let sbom_v1 = system
-        .ingest_sbom(
-            ("source", "http://sbom.com/test.json"),
-            &Digests::digest("8"),
-            "a",
-            (),
-            Transactional::None,
-        )
-        .await?;
-    let sbom_v2 = system
-        .ingest_sbom(
-            ("source", "http://sbom.com/test.json"),
-            &Digests::digest("9"),
-            "b",
-            (),
-            Transactional::None,
-        )
-        .await?;
-    let sbom_v3 = system
-        .ingest_sbom(
-            ("source", "http://sbom.com/test.json"),
-            &Digests::digest("10"),
-            "c",
-            (),
-            Transactional::None,
-        )
-        .await?;
-
-    sbom_v1
-        .ingest_describes_cpe22(
-            "cpe:/a:redhat:quarkus:2.13::el8".parse()?,
-            Transactional::None,
-        )
-        .await?;
-
-    sbom_v2
-        .ingest_describes_cpe22(
-            "cpe:/a:redhat:quarkus:2.13::el8".parse()?,
-            Transactional::None,
-        )
-        .await?;
-
-    sbom_v3
-        .ingest_describes_cpe22(
-            "cpe:/a:redhat:not-quarkus:2.13::el8".parse()?,
-            Transactional::None,
-        )
-        .await?;
-
-    let found = system
-        .locate_sboms(
-            SbomLocator::Cpe("cpe:/a:redhat:quarkus:2.13::el8".parse()?),
-            Transactional::None,
-        )
-        .await?;
-
-    assert_eq!(2, found.len());
-    assert!(found.contains(&sbom_v1));
-    assert!(found.contains(&sbom_v2));
-
-    Ok(())
-}
-
-#[test_context(TrustifyContext, skip_teardown)]
-#[test(tokio::test)]
-async fn transitive_dependency_of(ctx: TrustifyContext) -> Result<(), anyhow::Error> {
-    let db = ctx.db;
-    let system = Graph::new(db);
-
-    let sbom1 = system
-        .ingest_sbom(
-            ("source", "http://sbomsRus.gov/thing1.json"),
-            &Digests::digest("8675309"),
-            "a",
-            (),
-            Transactional::None,
-        )
-        .await?;
-
-    sbom1
-        .ingest_package_relates_to_package(
-            Purl::from_str("pkg://maven/io.quarkus/transitive-b@1.2.3")?,
-            Relationship::DependencyOf,
-            Purl::from_str("pkg://maven/io.quarkus/transitive-a@1.2.3")?,
-            Transactional::None,
-        )
-        .await?;
-
-    sbom1
-        .ingest_package_relates_to_package(
-            Purl::from_str("pkg://maven/io.quarkus/transitive-c@1.2.3")?,
-            Relationship::DependencyOf,
-            Purl::from_str("pkg://maven/io.quarkus/transitive-b@1.2.3")?,
-            Transactional::None,
-        )
-        .await?;
-
-    sbom1
-        .ingest_package_relates_to_package(
-            Purl::from_str("pkg://maven/io.quarkus/transitive-d@1.2.3")?,
-            Relationship::DependencyOf,
-            Purl::from_str("pkg://maven/io.quarkus/transitive-c@1.2.3")?,
-            Transactional::None,
-        )
-        .await?;
-
-    sbom1
-        .ingest_package_relates_to_package(
-            Purl::from_str("pkg://maven/io.quarkus/transitive-e@1.2.3")?,
-            Relationship::DependencyOf,
-            Purl::from_str("pkg://maven/io.quarkus/transitive-c@1.2.3")?,
-            Transactional::None,
-        )
-        .await?;
-
-    sbom1
-        .ingest_package_relates_to_package(
-            Purl::from_str("pkg://maven/io.quarkus/transitive-d@1.2.3")?,
-            Relationship::DependencyOf,
-            Purl::from_str("pkg://maven/io.quarkus/transitive-b@1.2.3")?,
-            Transactional::None,
-        )
-        .await?;
-
-    let _results = sbom1
-        .related_packages_transitively(
-            &[Relationship::DependencyOf],
-            &"pkg://maven/io.quarkus/transitive-a@1.2.3".try_into()?,
-            Transactional::None,
-        )
-        .await?;
-
-    Ok(())
-}
-
-#[test_context(TrustifyContext, skip_teardown)]
-#[test(tokio::test)]
-async fn ingest_package_relates_to_package_dependency_of(
-    ctx: TrustifyContext,
-) -> Result<(), anyhow::Error> {
-    let db = ctx.db;
-    let system = Graph::new(db.clone());
-    let fetch = SbomService::new(db);
-
-    let sbom1 = system
-        .ingest_sbom(
-            ("source", "http://sbomsRus.gov/thing1.json"),
-            &Digests::digest("8675309"),
-            "a",
-            (),
-            Transactional::None,
-        )
-        .await?;
-
-    sbom1
-        .ingest_package_relates_to_package(
-            Purl::from_str("pkg://maven/io.quarkus/quarkus-postgres@1.2.3")?,
-            Relationship::DependencyOf,
-            Purl::from_str("pkg://maven/io.quarkus/quarkus-core@1.2.3")?,
-            Transactional::None,
-        )
-        .await?;
-
-    let sbom2 = system
-        .ingest_sbom(
-            ("source", "http://sbomsRus.gov/thing2.json"),
-            &Digests::digest("8675308"),
-            "b",
-            (),
-            Transactional::None,
-        )
-        .await?;
-
-    sbom2
-        .ingest_package_relates_to_package(
-            Purl::from_str("pkg://maven/io.quarkus/quarkus-sqlite@1.2.3")?,
-            Relationship::DependencyOf,
-            Purl::from_str("pkg://maven/io.quarkus/quarkus-core@1.2.3")?,
-            Transactional::None,
-        )
-        .await?;
-
-    let dependencies = fetch
-        .related_packages(
-            sbom1.sbom.sbom_id,
-            Relationship::DependencyOf,
-            "pkg://maven/io.quarkus/quarkus-core@1.2.3",
-            Transactional::None,
-        )
-        .await?;
-
-    assert_eq!(1, dependencies.len());
-
-    assert_eq!(
-        vec!["pkg://maven/io.quarkus/quarkus-postgres@1.2.3"],
-        dependencies[0].purl
-    );
-
-    let dependencies = fetch
-        .related_packages(
-            sbom2.sbom.sbom_id,
-            Relationship::DependencyOf,
-            "pkg://maven/io.quarkus/quarkus-core@1.2.3",
-            Transactional::None,
-        )
-        .await?;
-
-    assert_eq!(1, dependencies.len());
-
-    assert_eq!(
-        vec!["pkg://maven/io.quarkus/quarkus-sqlite@1.2.3"],
-        dependencies[0].purl
-    );
-
-    Ok(())
-}
-
-#[test_context(TrustifyContext, skip_teardown)]
-#[test(tokio::test)]
-async fn sbom_vulnerabilities(ctx: TrustifyContext) -> Result<(), anyhow::Error> {
-    let db = ctx.db;
-    let system = Graph::new(db);
-
-    log::debug!("{:?}", system);
-
-    let sbom = system
-        .ingest_sbom(
-            ("source", "http://sbomsRus.gov/thing1.json"),
-            &Digests::digest("8675309"),
-            "a",
-            (),
-            Transactional::None,
-        )
-        .await?;
-
-    log::debug!("-------------------- A");
-
-    sbom.ingest_describes_package("pkg://oci/my-app@1.2.3".try_into()?, Transactional::None)
-        .await?;
-    log::debug!("-------------------- B");
-
-    sbom.ingest_package_relates_to_package(
-        Purl::from_str("pkg://maven/io.quarkus/quarkus-core@1.2.3")?,
-        Relationship::DependencyOf,
-        Purl::from_str("pkg://oci/my-app@1.2.3")?,
-        Transactional::None,
-    )
-    .await?;
-    log::debug!("-------------------- C");
-
-    sbom.ingest_package_relates_to_package(
-        Purl::from_str("pkg://maven/io.quarkus/quarkus-postgres@1.2.3")?,
-        Relationship::DependencyOf,
-        Purl::from_str("pkg://maven/io.quarkus/quarkus-core@1.2.3")?,
-        Transactional::None,
-    )
-    .await?;
-    log::debug!("-------------------- D");
-
-    sbom.ingest_package_relates_to_package(
-        Purl::from_str("pkg://maven/postgres/postgres-driver@1.2.3")?,
-        Relationship::DependencyOf,
-        Purl::from_str("pkg://maven/io.quarkus/quarkus-postgres@1.2.3")?,
-        Transactional::None,
-    )
-    .await?;
-
-    let advisory = system
-        .ingest_advisory(
-            "RHSA-1",
-            ("source", "http://redhat.com/secdata/RHSA-1"),
-            &Digests::digest("7"),
-            (),
-            Transactional::None,
-        )
-        .await?;
-
-    let _advisory_vulnerability = advisory
-        .link_to_vulnerability("CVE-00000001", None, Transactional::None)
-        .await?;
-
-    Ok(())
-}
-
-/*
-#[tokio::test]
-async fn ingest_contains_packages() -> Result<(), anyhow::Error> {
-    env_logger::builder()
-    .filter_level(log::LevelFilter::Info)
-    .is_test(true)
-    .init();
-
-        let fetch = InnerSystem::for_test("ingest_contains_packages").await?;
-
-        let sbom = fetch
-            .ingest_sbom("http://sboms.mobi/something.json", "7", Transactional::None)
-            .await?;
-
-        let contains1 = sbom
-            .ingest_contains_package(
-                "pkg://maven/io.quarkus/quarkus-core@1.2.3",
-                Transactional::None,
-            )
-            .await?;
-
-        let contains2 = sbom
-            .ingest_contains_package(
-                "pkg://maven/io.quarkus/quarkus-core@1.2.3",
-                Transactional::None,
-            )
-            .await?;
-
-        let contains3 = sbom
-            .ingest_contains_package(
-                "pkg://maven/io.quarkus/quarkus-addons@1.2.3",
-                Transactional::None,
-            )
-            .await?;
-
-        assert_eq!(
-            contains1.sbom_contains_package.qualified_package_id,
-            contains2.sbom_contains_package.qualified_package_id
-        );
-        assert_ne!(
-            contains1.sbom_contains_package.qualified_package_id,
-            contains3.sbom_contains_package.qualified_package_id
-        );
-
-        let mut contains = sbom.contains_packages(Transactional::None).await?;
-
-        assert_eq!(2, contains.len());
-
-        let contains: Vec<_> = contains.drain(0..).map(Purl::from).collect();
-
-        assert!(contains.contains(&Purl::from("pkg://maven/io.quarkus/quarkus-core@1.2.3")));
-        assert!(contains.contains(&Purl::from("pkg://maven/io.quarkus/quarkus-addons@1.2.3")));
-
-        Ok(())
-    }
-     */
-
-/*
-
-#[tokio::test]
-async fn ingest_and_fetch_sbom_packages() -> Result<(), anyhow::Error> {
-    /*
-    env_logger::builder()
-        .filter_level(log::LevelFilter::Info)
-        .is_test(true)
-        .init();
-
-     */
-    let fetch = InnerSystem::for_test("ingest_and_fetch_sbom_packages").await?;
-
-    let sbom_v1 = fetch.ingest_sbom("http://sbom.com/test.json", "8").await?;
-    let sbom_v2 = fetch.ingest_sbom("http://sbom.com/test.json", "9").await?;
-    let sbom_v3 = fetch
-        .ingest_sbom("http://sbom.com/test.json", "10")
-        .await?;
-
-    sbom_v1
-        .ingest_sbom_dependency("pkg://maven/io.quarkus/taco@1.2.3", Transactional::None)
-        .await?;
-
-    sbom_v1
-        .ingest_package_dependency(
-            "pkg://maven/io.quarkus/foo@1.2.3",
-            "pkg://maven/io.quarkus/baz@1.2.3",
-            Transactional::None,
-        )
-        .await?;
-
-    sbom_v2
-        .ingest_package_dependency(
-            "pkg://maven/io.quarkus/foo@1.2.3",
-            "pkg://maven/io.quarkus/bar@1.2.3",
-            Transactional::None,
-        )
-        .await?;
-
-    let sbom_packages = sbom_v1.all_packages(Transactional::None).await?;
-    assert_eq!(3, sbom_packages.len());
-
-    for sbom_package in sbom_packages {
-        let _sboms = sbom_package
-            .package
-            .sboms_containing(Transactional::None)
-            .await?;
-    }
-
-    Ok(())
-}
-
- */
